Fix test timeouts and other failures.
